python/tutorials/tour-of-workflows-python/pyproject.toml
[project]
name = "tour-of-workflows-python"
version = "0.1.0"
description = "Restate SDK Tour of Workflows"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hypercorn>=0.17.3",
    "pydantic>=2.10.6",
    "requests>=2.32.5",
    "restate_sdk[serde]>=0.9.1",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
